# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/swagger":{"get":{"operationId":"AppController_swagger","parameters":[],"responses":{"200":{"description":""}}}},"/api/members":{"get":{"operationId":"MembersController_getMembers","parameters":[{"name":"page","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":1,"type":"string"}},{"name":"perPage","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":10,"type":"string"}},{"name":"names","required":false,"in":"query","description":"parametro filtros","schema":{"type":"string"}},{"name":"lastnames","required":false,"in":"query","description":"parametro filtros","schema":{"type":"string"}},{"name":"email","required":false,"in":"query","description":"parametro filtros","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["members"]},"post":{"operationId":"MembersController_createMember","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberBodyDto"}}}},"responses":{"201":{"description":""}},"tags":["members"]}},"/api/members/{id}":{"get":{"operationId":"MembersController_getMemberById","parameters":[{"name":"id","required":true,"in":"path","description":"id del miembro","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["members"]},"put":{"operationId":"MembersController_updateMember","parameters":[{"name":"id","required":true,"in":"path","description":"id del miembro","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberBodyDto"}}}},"responses":{"200":{"description":""}},"tags":["members"]},"delete":{"operationId":"MembersController_deleteMember","parameters":[{"name":"id","required":true,"in":"path","description":"id del miembro","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["members"]}},"/api/selectors/municipalities":{"get":{"operationId":"SelectorsController_getMunicipalities","parameters":[{"name":"page","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":1,"type":"string"}},{"name":"perPage","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":10,"type":"string"}},{"name":"name","required":false,"in":"query","description":"parametro filtros","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["selectors"]}},"/api/selectors/occupations":{"get":{"operationId":"SelectorsController_getOccupations","parameters":[{"name":"page","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":1,"type":"string"}},{"name":"perPage","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":10,"type":"string"}},{"name":"name","required":false,"in":"query","description":"parametro filtros","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["selectors"]}},"/api/selectors/social-networks":{"get":{"operationId":"SelectorsController_getSocialNetworks","parameters":[{"name":"page","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":1,"type":"string"}},{"name":"perPage","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":10,"type":"string"}},{"name":"name","required":false,"in":"query","description":"parametro filtros","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["selectors"]}},"/api/selectors/how-know":{"get":{"operationId":"SelectorsController_getHowKnow","parameters":[{"name":"page","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":1,"type":"string"}},{"name":"perPage","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":10,"type":"string"}},{"name":"name","required":false,"in":"query","description":"parametro filtros","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["selectors"]}},"/api/selectors/steps":{"get":{"operationId":"SelectorsController_getStep","parameters":[{"name":"page","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":1,"type":"string"}},{"name":"perPage","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":10,"type":"string"}},{"name":"name","required":false,"in":"query","description":"parametro filtros","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["selectors"]}},"/api/auth":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthBodyDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/courses":{"get":{"operationId":"CoursesController_getCourses","parameters":[{"name":"page","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":1,"type":"string"}},{"name":"perPage","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":10,"type":"string"}},{"name":"name","required":false,"in":"query","description":"parametro filtros","schema":{"type":"string"}},{"name":"active","required":false,"in":"query","description":"parametro filtros","schema":{"type":"boolean"}},{"name":"startDate","required":false,"in":"query","description":"parametro filtros","schema":{"format":"date-time","type":"string"}},{"name":"endDate","required":false,"in":"query","description":"parametro filtros","schema":{"format":"date-time","type":"string"}}],"responses":{"200":{"description":""}},"tags":["courses"]},"post":{"operationId":"CoursesController_createCourse","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseBodyDto"}}}},"responses":{"201":{"description":""}},"tags":["courses"]}},"/api/courses/{id}":{"get":{"operationId":"CoursesController_getCourseById","parameters":[{"name":"id","required":true,"in":"path","description":"id del curso","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["courses"]},"put":{"operationId":"CoursesController_updateCourse","parameters":[{"name":"id","required":true,"in":"path","description":"id del curso","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseBodyDto"}}}},"responses":{"200":{"description":""}},"tags":["courses"]},"delete":{"operationId":"CoursesController_deleteCourse","parameters":[{"name":"id","required":true,"in":"path","description":"id del curso","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["courses"]}},"/api/enrollmentCourses":{"get":{"operationId":"EnrollmentCoursesController_getEnrollmentCourses","parameters":[{"name":"page","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":1,"type":"string"}},{"name":"perPage","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":10,"type":"string"}},{"name":"state","required":false,"in":"query","description":"filtro estado del curso","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["enrollmentCourses"]},"post":{"operationId":"EnrollmentCoursesController_createEnrollmentCourse","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrollmentCourseBodyDto"}}}},"responses":{"201":{"description":""}},"tags":["enrollmentCourses"]}},"/api/enrollmentCourses/{id}":{"get":{"operationId":"EnrollmentCoursesController_getEnrollmentCourseById","parameters":[{"name":"id","required":true,"in":"path","description":"id del curso","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["enrollmentCourses"]},"put":{"operationId":"EnrollmentCoursesController_updateEnrollmentCourse","parameters":[{"name":"id","required":true,"in":"path","description":"id del curso","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrollmentCourseBodyDto"}}}},"responses":{"200":{"description":""}},"tags":["enrollmentCourses"]},"delete":{"operationId":"EnrollmentCoursesController_deleteEnrollmentCourse","parameters":[{"name":"id","required":true,"in":"path","description":"id del curso","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["enrollmentCourses"]}},"/api/enrollmentCourses/byCourseId/{courseId}":{"get":{"operationId":"EnrollmentCoursesController_getEnrollmentCourseByCourseId","parameters":[{"name":"courseId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["enrollmentCourses"]}},"/api/enrollmentCourses/byMemberId/{memberId}":{"get":{"operationId":"EnrollmentCoursesController_getEnrollmentCourseByMemberId","parameters":[{"name":"memberId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["enrollmentCourses"]}},"/api/coursesClasses":{"get":{"operationId":"CoursesClassesController_getCoursesClasses","parameters":[{"name":"page","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":1,"type":"string"}},{"name":"perPage","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":10,"type":"string"}},{"name":"classDate","required":false,"in":"query","description":"filtro por el Id de la escuela ","schema":{"format":"date-time","type":"string"}}],"responses":{"200":{"description":""}},"tags":["coursesClasses"]},"post":{"operationId":"CoursesClassesController_createCoursesClasses","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoursesClassesBodyDto"}}}},"responses":{"201":{"description":""}},"tags":["coursesClasses"]}},"/api/coursesClasses/{id}":{"get":{"operationId":"CoursesClassesController_getCoursesClassesById","parameters":[{"name":"id","required":true,"in":"path","description":"id de la clase","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["coursesClasses"]},"put":{"operationId":"CoursesClassesController_updateCoursesClasses","parameters":[{"name":"id","required":true,"in":"path","description":"id de la clase","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoursesClassesBodyDto"}}}},"responses":{"200":{"description":""}},"tags":["coursesClasses"]},"delete":{"operationId":"CoursesClassesController_deleteCoursesClasses","parameters":[{"name":"id","required":true,"in":"path","description":"id de la clase","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["coursesClasses"]}},"/api/coursesClasses/byCourseId/{courseId}":{"get":{"operationId":"CoursesClassesController_getCoursesClassesByCourseId","parameters":[{"name":"courseId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["coursesClasses"]}},"/api/stepTopics":{"get":{"operationId":"StepTopicsController_getStepTopics","parameters":[{"name":"page","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":1,"type":"string"}},{"name":"perPage","required":false,"in":"query","description":"parametro para paginacion","schema":{"default":10,"type":"string"}},{"name":"topicName","required":false,"in":"query","description":"filtro por el nombre del tema ","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["stepTopics"]},"post":{"operationId":"StepTopicsController_createStepTopics","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StepTopicsBodyDto"}}}},"responses":{"201":{"description":""}},"tags":["stepTopics"]}},"/api/stepTopics/{id}":{"get":{"operationId":"StepTopicsController_getStepTopicsById","parameters":[{"name":"id","required":true,"in":"path","description":"id del tema","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["stepTopics"]},"put":{"operationId":"StepTopicsController_updateStepTopics","parameters":[{"name":"id","required":true,"in":"path","description":"id del tema","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StepTopicsBodyDto"}}}},"responses":{"200":{"description":""}},"tags":["stepTopics"]},"delete":{"operationId":"StepTopicsController_deleteStepTopics","parameters":[{"name":"id","required":true,"in":"path","description":"id del tema","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["stepTopics"]}},"/api/stepTopics/byStepId/{stepId}":{"get":{"operationId":"StepTopicsController_getStepTopicsByStepId","parameters":[{"name":"stepId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["stepTopics"]}},"/api/stepTopics/byCourseId/{courseId}":{"get":{"operationId":"StepTopicsController_getStepTopicsByCourseId","parameters":[{"name":"courseId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["stepTopics"]}}},"info":{"title":"The Church Medellin","description":"The Church Medellin","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"https://serverlessmicasamed.azurewebsites.net"},{"url":"http://localhost:3000"}],"components":{"schemas":{"MemberBodyDto":{"type":"object","properties":{"names":{"type":"string","description":"parametro de miembro"},"lastnames":{"type":"string","description":"parametro de miembro"},"age":{"type":"number","description":"parametro de miembro"},"sex":{"type":"string","description":"parametro de miembro","enum":["M","F","O"]},"phone":{"type":"string","description":"parametro de miembro"},"email":{"type":"string","description":"parametro de miembro"},"district":{"type":"string","description":"parametro de miembro"},"volunteer":{"type":"boolean","description":"parametro de miembro"},"municipalityId":{"type":"number","description":"parametro de miembro"},"occupationId":{"type":"number","description":"parametro de miembro"},"socialNetworkId":{"type":"number","description":"parametro de miembro"},"howKnowId":{"type":"number","description":"parametro de miembro"},"arrivalDate":{"format":"date-time","type":"string","description":"parametro de miembro"},"comment":{"type":"string","description":"parametro de miembro"}},"required":["names","lastnames","age","sex","phone","district","volunteer","municipalityId","occupationId","socialNetworkId","howKnowId","arrivalDate","comment"]},"AuthBodyDto":{"type":"object","properties":{"email":{"type":"string","description":"parametro de auth"},"password":{"type":"string","description":"parametro de auth"}},"required":["email","password"]},"CourseBodyDto":{"type":"object","properties":{"name":{"type":"string","description":"parametro de curso - escuela"},"stepId":{"type":"number","description":"parametro de curso - escuela"},"active":{"type":"boolean","description":"parametro de curso - escuela"},"startDate":{"format":"date-time","type":"string","description":"parametro de curso - escuela"},"endDate":{"format":"date-time","type":"string","description":"parametro de curso - escuela"}},"required":["name","stepId","active","startDate","endDate"]},"EnrollmentCourseBodyDto":{"type":"object","properties":{"state":{"type":"string","description":"parametro de la inscripcion del curso - escuela"},"memberId":{"type":"number","description":"parametro de la inscripcion del curso - escuela"},"courseId":{"type":"number","description":"parametro de la inscripcion del curso - escuela"}},"required":["state","memberId","courseId"]},"CoursesClassesBodyDto":{"type":"object","properties":{"numberClass":{"type":"number","description":"parametro de la creación de la clase"},"classDate":{"format":"date-time","type":"string","description":"parametro de la creación de la clase"},"topicId":{"type":"number","description":"parametro de la creación de la clase"},"courseId":{"type":"number","description":"parametro de la creación de la clase"}},"required":["numberClass","classDate","topicId","courseId"]},"StepTopicsBodyDto":{"type":"object","properties":{"topicName":{"type":"string","description":"parametro de la creación del tema"},"stepId":{"type":"number","description":"parametro de la creación del tema"}},"required":["topicName","stepId"]}}}}